Removed Level/Sprite Color System
It was a disaster.

In an attempt to not have to create new sprites with new color schemes I created a system parallel to my level system which applied a color to the sprite renderers. The problem is that I could only add to the colors that were already there. They could only go so far before becoming hideous, not to mention becoming unusable for people with red/green color blindness.